Compile go from source for cross compilation
This takes a little (lot) longer, but is necessary if we want to cross
compile for other platforms (we do).

Initially I had a call to goxc -t which builds the necessary tooling to
cross compile for those other platforms, but that takes like 10 minutes
and it gets handled automagically the first time you cross compile
anyway.